Add files via upload
The following code solves for the concentration profile of hydrocarbons as 
along the z direction of a non isothermal tubular reactor. The reactor is 95 
meters long with a diameter of 0.1 meters. The product of interest is ethylene.
The code simulataneous solves 9 fluxes ODEs, 1 pressure ODE and 1 temperature ODE
and then updates the  velocity using RK4. Functions were created to calculate 
specific heats, heats of reactions, thermal conductivities, overall heat transfer
and viscosity. Parameter such as the wall temperature, length of reactor, and feed
can be varied to increase the selectivity of ethylene and decrease the conversion
of ethane. If the conversion of ethane is above 0.6 the reactor will coke up.